requirements/prod.txt
aiogram==2.14.3
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
babel==2.9.1
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.1
idna==3.2
jinja2==3.0.1
markupsafe==2.0.1
multidict==5.1.0
psycopg2-binary==2.9.1
pytz==2021.1
redis==3.5.3
requests==2.26.0
soupsieve==2.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
aiopg==1.1.0
peewee-async
peewee_migrate2